requirements.txt
# Core dependencies
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.3.0

# Configuration
pyyaml>=6.0

# Data sources (public)
yfinance>=0.2.0
requests>=2.31.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Statistical modeling
statsmodels>=0.14.0

# Excel support
openpyxl>=3.1.0
